Read description before merge!!
Flight API does not work and there is no free alternative for what we are trying to build.

This is the design if the group decides to go with a rewrite.

I have saved all of the old code in .old files. A lot was reused in this build.

DO NOT MERGE WITHOUT MAJORITY APPROVAL.